Separated some of CorpusBuilder into KeySearch class
Limited the CorpusBuilder class to just parsing the text and building
the dictionary and tf-idf model. Further operations on the constructed
corpus are done with KeySearch.
