+++ corpusbuilder.py
@@ -6,11 +6,14 @@ Created on Mon Oct 17 21:19:15 2016
 """
 
 import nltk
-import textwrap
+import re
 from gensim import corpora
 from gensim.models import TfidfModel
 from collections import defaultdict
-import pickle
+from keysearch import KeySearch
+from os import listdir, makedirs
+from os.path import isfile, join, exists
+
 
 # I lazily made this a global constant so that I wouldn't have to include
 # it in the save and load features.
@@ -86,7 +89,7 @@ class CorpusBuilder(object):
     form.
         
     """
-    def __init__(self):
+    def __init__(self, stop_words_file='', doc_start_pattern='', doc_start_is_separator=True, sub_patterns=[]):
         """
         `stop_words_file` is the path and filename to a file containing stop
         words to be filtered from the input text. There should be one token
@@ -99,8 +102,8 @@ class CorpusBuilder(object):
         self.titles = []        
         
         # Create mappings for the entry tags.
-        self.tagsToEntries = {}
-        self.entriesToTags = []
+        self.tagsToDocs = {}
+        self.docsToTags = []
 
         self.files = []
         self.doc_line_nums = []
@@ -109,15 +112,11 @@ class CorpusBuilder(object):
         # This is a temporary data structure used for filtering out words
         # that only occur once in the corpus.
         # The final word counts can be found in self.dictionary        
-        self.frequency = defaultdict(int)               
-        
-    def setStopWordList(self, stop_words_file):
-        """
-        Specify the list of "stop words" to be removed from the corpus. 
-        """
+        self.frequency = defaultdict(int)
+
         # Read in all of the stop words (one per line) and store them as a set.
         # The call to f.read().splitlines() reads the lines without the newline
-        # char.       
+        # char.           
         with open(stop_words_file) as f:
             lines = f.read().splitlines()
             
@@ -128,7 +127,27 @@ class CorpusBuilder(object):
                 stoplist.append(line.decode(enc_format))
                 
             # Convert to a set representation.    
-            self.stoplist = set(stoplist)
+            self.stoplist = set(stoplist)               
+        
+        # Record the regex pattern for the start of a document.
+        self.doc_start_pattern = doc_start_pattern
+
+        # Record whether the doc start pattern is the first line of the doc
+        # or just a separator (for example, an empty line)
+        self.doc_start_is_separator = doc_start_is_separator        
+        
+        # Record any regex substitutions to make.
+        self.sub_patterns = sub_patterns
+    
+    def applyRegExFilters(self, line):
+        """
+        Remove tokens matching some regex filters.
+        """
+
+        for substitution in self.sub_patterns:                    
+            line = re.sub(substitution[0], substitution[1], line)
+            
+        return line    
     
     def addDocument(self, title, lines, tags=[], filename=None, doc_start=None, doc_end=None):
         """
@@ -161,8 +180,8 @@ class CorpusBuilder(object):
         # Store the title.
         self.titles.append(title)
 
-        # Store the list of tags for this journal entry.
-        self.entriesToTags.append(tags)                         
+        # Store the list of tags for this doc.
+        self.docsToTags.append(tags)                         
             
         # Add mappings from the tags to this journal entry.
         for tag in tags:
@@ -170,10 +189,10 @@ class CorpusBuilder(object):
             tag = tag.lower()
                 
             # Add the tag to the dictionary.
-            if tag in self.tagsToEntries:
-                self.tagsToEntries[tag].append(docID)
+            if tag in self.tagsToDocs:
+                self.tagsToDocs[tag].append(docID)
             else:
-                self.tagsToEntries[tag] = [docID]   
+                self.tagsToDocs[tag] = [docID]   
 
         # Store the filenames only once.
         if filename in self.files:        
@@ -220,6 +239,92 @@ class CorpusBuilder(object):
         # Add this document to the list of all documents.
         self.documents.append(doc)
     
+    
+    def addFile(self, filepath, filename):
+        """
+        
+        """
+        # Read in the text file
+        with open(filepath) as f:
+            content = f.readlines()
+
+        # TODO! Add tag support.
+        doc_tags = []    
+        
+        doc_title = ""       
+        doc = []
+        doc_start = -1
+        
+        # For each line in the file...
+        for lineNum in range(0, len(content)):
+
+            # Get the next line.
+            line = content[lineNum]                        
+                        
+            # Check for the pattern matching the start of a document.
+            matchStart = re.search(self.doc_start_pattern, line)                        
+
+            # Whether to add this line to the current document.
+            add_line = True            
+            
+            # If we've found the start of a new doc...            
+            if matchStart and doc:                
+                
+                # Note the line number of the last line of the current document.
+                doc_end = lineNum - 1 + 1                
+                
+                # Only add entries that are longer than one line.
+                # This filters out section headings.                
+                # TODO - Make this configurable.
+                if len(doc) > 1:
+                    self.addDocument(title=doc_title, lines=doc, tags=doc_tags, filename=filepath,doc_start=doc_start, doc_end=doc_end)
+                
+                doc = []
+                doc_title = ""
+                doc_start = -1
+
+                # If the doc_start_pattern is a separator, don't add it to
+                # the new doc (start the doc on the next line).            
+                if self.doc_start_is_separator:
+                    add_line = False
+                
+            # Add this line to the current doc.
+            if add_line:
+                # If this doc doesn't have a title yet, use this line.                
+                if not doc_title:                
+                    doc_title = filename + ' - ' + line       
+                    doc_start = lineNum + 1 # Line numbers are numbered from 1.
+                
+                # Run reg ex filters to remove some tokens.
+                line = self.applyRegExFilters(line)                
+                
+                # Add the line to the doc.
+                doc.append(line)
+        
+        # End of file reached.
+        doc_end = lineNum - 1 + 1;        
+        self.addDocument(doc_title, doc, doc_tags, filename=filepath ,doc_start=doc_start, doc_end=doc_end)
+   
+    def addDirectory(self, dir_path):
+        """
+        Add all of the .txt files in the specified directory to the corpus.
+        """
+        # Get all the source text files.
+        files = [f for f in listdir(dir_path) if isfile(join(dir_path, f))]
+
+        # For each file in the directory:
+        for f in files:
+
+            # Only parse the .txt files.    
+            if not (f[-4:] == '.txt'):
+                continue   
+    
+            print '  Parsing file:', f
+            
+            # Pass down to addFile.
+            self.addFile(filepath=(dir_path + f), filename=f[0:-4])
+            
+
    
     def buildCorpus(self):
         """
@@ -247,300 +352,9 @@ class CorpusBuilder(object):
         self.tfidf_model = TfidfModel(corpus)
         self.corpus_tfidf = self.tfidf_model[corpus]
     
-    
-    def getTfidfForText(self, text):
-        """
-        This function takes new input `text` (not part of the original corpus),
-        and processes it into a tf-idf vector.
+    def toKeySearch(self):
+        ksearch = KeySearch(self.dictionary, self.tfidf_model, 
+                            self.corpus_tfidf, self.titles, self.tagsToDocs,
+                            self.docsToTags, self.files, self.doc_line_nums)        
         
-        The input text should be a single string.
-        """
-        # If the string is not already unicode, decode the string into unicode
-        # so the NLTK can handle it.
-        if isinstance(text, str):
-            try:    
-                text = text.decode(enc_format)        
-            except:
-                print '======== Failed to decode input text! ========'
-                print 'Make sure text is encoded in', enc_format
-                print 'Input text:'
-                print text
-                return []
-        
-        # If the string ends in a newline, remove it.
-        text = text.replace('\n', ' ')
-
-        # Convert everything to lowercase, then use NLTK to tokenize.
-        tokens = nltk.word_tokenize(text.lower())
-       
-        # Remove stop words.
-        # TODO - I think I shouldn't need to do this, it was only for building
-        #       the dictionary. 
-        #tokens = [word for word in tokens if word not in self.stoplist]
-
-        # Convert the tokenized text into a bag of words representation.
-        bow_vec = self.dictionary.doc2bow(tokens) 
-        
-        # Convert the bag-of-words representation to tf-idf
-        return self.tfidf_model[bow_vec]
-    
-    def getTfidfForFile(self, filename):
-        """
-        Convert the text in the provided file to a tf-idf vector.
-        """
-        # Open the file and read all lines.        
-        with open(filename) as f:
-            text = f.readlines()
-
-        # Combine the lines into a single string.
-        text = " ".join(text)
-
-        # Pass the text down.
-        return self.getTfidfForText(text)
-    
-    def getTfidfForDoc(self, doc_id):
-        """
-        Return the tf-idf vector for the specified document.
-        """        
-        return self.corpus_tfidf[doc_id]
-
-    def keywordSearch(self, includes=[], excludes=[], docs=[]):
-        """
-        Performs a boolean keyword search over the corpus.
-        
-        All words in the dictionary are lower case. This function will convert
-        all supplied keywords to lower case.
-        
-        Parameters:
-            includes    A list of words (as strings) that the documents 
-                        *must include*.
-            excludes    A list of words (as strings) that the documents
-                        *must not include*.
-            docs        The list of documents to search in, represented by
-                        by doc_ids. If this list is empty, the entire corpus
-                        is searched.
-        """
-        
-        # If no doc ids were supplied, search the entire corpus.
-        if not docs:
-            docs = range(0, len(self.corpus_tfidf))
-    
-        # Convert all the keywords to their IDs.
-        # Force them to lower case in the process.
-        include_ids = []
-        exclude_ids = []
-    
-        for word in includes:
-            # Lookup the ID for the word.            
-            word_id = self.getIDForWord(word.lower())            
-            
-            # Verify the word exists in the dictionary.
-            if word_id == -1:
-                raise 'Word', word.lower(), 'not in dictionary!'
-            
-            # Add the word id to the list.
-            include_ids.append(word_id)
-            
-        for word in excludes:
-            exclude_ids.append(self.getIDForWord(word.lower()))
-        
-        results = []
-    
-        # For each of the documents to search...
-        
-        for doc_id in docs:
-            # Get the sparse tf-idf vector for the next document.
-            vec_tfidf = self.corpus_tfidf[doc_id]
-            
-            # Create a list of the word ids in this document.
-            doc_words = [tfidf[0] for tfidf in vec_tfidf]
-            
-            match = True
-            
-            # Check for words that must be present.
-            for word_id in include_ids:
-                if not word_id in doc_words:
-                    match = False
-                    break
-            
-            # If we failed the 'includes' test, skip to the next document.
-            if not match:
-                continue
-    
-            # Check for words that must not be present.
-            for word_id in exclude_ids:
-                if word_id in doc_words:
-                    match = False
-                    break
-            
-            # If we passed the 'excludes' test, this is a valid result.
-            if match:
-                results.append(doc_id)
-        
-        return results
-            
-    
-    def printTopNWords(self, topn=10):
-        """
-        Print the 'topn' most frequent words in the corpus.
-        
-        This is useful for checking to see if you have any common, bogus tokens
-        that need to be filtered out of the corpus.
-        """
-        
-        # Get the dictionary as a list of tuples.
-        # The tuple is (word_id, count)
-        word_counts = [(key, value) for (key, value) in self.dictionary.dfs.iteritems()]
-        
-        # Sort the list by the 'value' of the tuple (incidence count) 
-        from operator import itemgetter
-        word_counts = sorted(word_counts, key=itemgetter(1))
-        
-        # Print the most common words.
-        # The list is sorted smallest to biggest, so...
-        print 'Top', topn, 'most frequent words'
-        for i in range(-1, -topn, -1):
-            print '  %s   %d' % (self.dictionary[word_counts[i][0]].ljust(10), word_counts[i][1])
-    
-    def getVocabSize(self):
-        """
-        Returns the number of unique words in the final vocabulary (after all
-        filtering).
-        """
-        return len(self.dictionary.keys())
-        
-    def getIDForWord(self, input_word):
-        """
-        Lookup the ID for a specific word.
-        
-        Returns -1 if the word isn't in the dictionary.
-        """
-        
-        # All words in dictionary are lower case.
-        input_word = input_word.lower()
-        
-        # Search through the words in the dictionary to find the input.
-        for id, word in self.dictionary.iteritems():
-            if word == input_word:
-                return id
-        
-        # If it wasn't found, return -1.
-        return -1
-       
-    def getDocLocation(self, doc_id):
-        """
-        Return the filename and line numbers that 'doc_id' came from.
-        """
-        line_nums = self.doc_line_nums[doc_id]        
-        filename = self.files[line_nums[0]]
-        return filename, line_nums[1], line_nums[2]
-    
-    def readDocSource(self, doc_id):
-        """
-        Reads the original source file for the document 'doc_id' and retrieves
-        the source lines.
-        """
-        # Lookup the source for the doc.
-        line_nums = self.doc_line_nums[doc_id]        
-        
-        filename = self.files[line_nums[0]]
-        line_start = line_nums[1]
-        line_end = line_nums[2]
-
-        results = []        
-
-        # Open the file and read just the specified lines.        
-        with open(filename) as fp:
-            for i, line in enumerate(fp):
-                # 'i' starts at 0 but line numbers start at 1.
-                line_num = i + 1
-                
-                if line_num > line_end:
-                    break
-                
-                if line_num >= line_start:
-                    results.append(line)
-    
-        return results
-    
-    def printDocSourcePretty(self, doc_id, max_lines=8, indent='    '):
-        """
-        Prints the original source lines for the document 'doc_id'.
-        
-        This function leverages the 'textwrap' Python module to limit the 
-        print output to 80 columns.        
-        """
-            
-        # Read in the document.
-        lines = self.readDocSource(doc_id)
-            
-        # Limit the result to 'max_lines'.
-        truncated = False
-        if len(lines) > max_lines:
-            truncated = True
-            lines = lines[0:max_lines]
-
-        # Convert the list of strings to a single string.
-        lines = '\n'.join(lines)
-
-        # Remove indentations in the source text.
-        dedented_text = textwrap.dedent(lines).strip()
-        
-        # Add an ellipsis to the end to show we truncated the doc.
-        if truncated:
-            dedented_text = dedented_text + ' ...'
-        
-        # Wrap the text so it prints nicely--within 80 columns.
-        # Print the text indented slightly.
-        pretty_text = textwrap.fill(dedented_text, initial_indent=indent, subsequent_indent=indent, width=80)
-        
-        print pretty_text   
-    
-    def save(self, save_dir='./'):
-        """
-        Write out the built corpus to a save directory.
-        """
-        # Store the tag tables.
-        pickle.dump((self.tagsToEntries, self.entriesToTags), open(save_dir + 'tag-tables.pickle', 'wb'))
-        
-        # Store the document titles.
-        pickle.dump(self.titles, open(save_dir + 'titles.pickle', 'wb'))
-        
-        # Write out the tfidf model.
-        self.tfidf_model.save(save_dir + 'documents.tfidf_model')
-        
-        # Write out the tfidf corpus.
-        corpora.MmCorpus.serialize(save_dir + 'documents_tfidf.mm', self.corpus_tfidf)  
-
-        # Write out the dictionary.
-        self.dictionary.save(save_dir + 'documents.dict')
-        
-        # Save the filenames.
-        pickle.dump(self.files, open(save_dir + 'files.pickle', 'wb'))
-        
-        # Save the file ID and line numbers for each document.
-        pickle.dump(self.doc_line_nums, open(save_dir + 'doc_line_nums.pickle', 'wb'))
-        
-        # Objects that are not saved:
-        #  - stop_list - You don't need to filter stop words for new input
-        #                text, they simply aren't found in the dictionary.
-        #  - frequency - This preliminary word count object is only used for
-        #                removing infrequent words. Final word counts are in
-        #                the `dictionary` object.
-        
-    def load(self, save_dir='./'):
-        """
-        Load the corpus from a save directory.
-        """
-        tables = pickle.load(open(save_dir + 'tag-tables.pickle', 'rb'))
-        self.tagsToEntries = tables[0]
-        self.entriesToTags = tables[1]        
-        self.titles = pickle.load(open(save_dir + 'titles.pickle', 'rb'))
-        self.tfidf_model = TfidfModel.load(fname=save_dir + 'documents.tfidf_model')
-        self.corpus_tfidf = corpora.MmCorpus(save_dir + 'documents_tfidf.mm')
-        self.dictionary = corpora.Dictionary.load(fname=save_dir + 'documents.dict')
-        self.files = pickle.load(open(save_dir + 'files.pickle', 'rb'))
-        self.doc_line_nums = pickle.load(open(save_dir + 'doc_line_nums.pickle', 'rb'))
-        
-        
-        
+        return ksearch
